Extract exclusive lock collection to its own class
