+++ timelock-impl/src/main/java/com/palantir/atlasdb/timelock/lock/AsyncLockService.java
@@ -155,12 +155,12 @@ public class AsyncLockService implements Closeable {
             Set<LockDescriptor> lockDescriptors,
             TimeLimit timeout,
             Optional<LockRequestMetadata> metadata) {
-        OrderedLocks orderedLocks = locks.getAll(lockDescriptors);
+        OrderedLocks orderedLocks = locks.getAllExclusive(lockDescriptors);
         return lockAcquirer.acquireLocks(requestId, orderedLocks, timeout, metadata);
     }
 
     private AsyncResult<Void> awaitLocks(UUID requestId, Set<LockDescriptor> lockDescriptors, TimeLimit timeout) {
-        OrderedLocks orderedLocks = locks.getAll(lockDescriptors);
+        OrderedLocks orderedLocks = locks.getAllExclusive(lockDescriptors);
         return lockAcquirer.waitForLocks(requestId, orderedLocks, timeout);
     }
